refactor: VoteService 기능을 PollService로 이전 및 관련 의존성 정리 (#490)
