+++ backend/src/test/java/com/tamnara/backend/poll/controller/PollControllerIntegrationTest.java
@@ -3,6 +3,7 @@ package com.tamnara.backend.poll.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tamnara.backend.global.util.CreateUserUtil;
 import com.tamnara.backend.poll.config.PollTestConfig;
+import com.tamnara.backend.poll.constant.PollResponseMessage;
 import com.tamnara.backend.poll.domain.Poll;
 import com.tamnara.backend.poll.domain.PollState;
 import com.tamnara.backend.poll.dto.request.PollCreateRequest;
@@ -11,7 +12,6 @@ import com.tamnara.backend.poll.dto.request.VoteRequest;
 import com.tamnara.backend.poll.dto.response.PollIdResponse;
 import com.tamnara.backend.poll.dto.response.PollInfoResponse;
 import com.tamnara.backend.poll.service.PollService;
-import com.tamnara.backend.poll.service.VoteService;
 import com.tamnara.backend.poll.util.PollCreateRequestTestBuilder;
 import com.tamnara.backend.user.domain.Role;
 import com.tamnara.backend.user.domain.State;
@@ -32,7 +32,6 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import java.util.List;
 
-import static com.tamnara.backend.poll.constant.PollResponseMessage.POLL_SCHEDULED;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.any;
@@ -55,7 +54,6 @@ class PollControllerIntegrationTest {
     @Autowired private ObjectMapper objectMapper;
 
     @Autowired private PollService pollService;
-    @Autowired private VoteService voteService;
 
     User user;
     @BeforeEach
@@ -111,7 +109,7 @@ class PollControllerIntegrationTest {
     void vote_success() throws Exception {
         // given
         VoteRequest voteRequest = new VoteRequest(List.of(1L, 2L));
-        given(voteService.vote(user, voteRequest)).willReturn(new PollIdResponse(1L));
+        given(pollService.vote(user, voteRequest)).willReturn(new PollIdResponse(1L));
 
         // when & then
 
@@ -134,6 +132,6 @@ class PollControllerIntegrationTest {
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.success").value(true))
-                .andExpect(jsonPath("$.message").value(POLL_SCHEDULED));
+                .andExpect(jsonPath("$.message").value(PollResponseMessage.POLL_SCHEDULED));
     }
 }
