+++ backend/src/test/java/com/tamnara/backend/poll/service/PollServiceImplTest.java
@@ -3,18 +3,20 @@ package com.tamnara.backend.poll.service;
 import com.tamnara.backend.alarm.constant.AlarmMessage;
 import com.tamnara.backend.alarm.domain.AlarmType;
 import com.tamnara.backend.alarm.event.AlarmEvent;
+import com.tamnara.backend.poll.constant.PollResponseMessage;
 import com.tamnara.backend.poll.domain.Poll;
 import com.tamnara.backend.poll.domain.PollOption;
 import com.tamnara.backend.poll.domain.PollState;
 import com.tamnara.backend.poll.dto.request.PollCreateRequest;
 import com.tamnara.backend.poll.dto.request.PollOptionCreateRequest;
+import com.tamnara.backend.poll.dto.request.VoteRequest;
 import com.tamnara.backend.poll.dto.response.PollInfoResponse;
 import com.tamnara.backend.poll.repository.PollOptionRepository;
 import com.tamnara.backend.poll.repository.PollRepository;
 import com.tamnara.backend.poll.repository.VoteRepository;
 import com.tamnara.backend.poll.util.PollCreateRequestTestBuilder;
-import com.tamnara.backend.poll.util.PollOptionTestBuilder;
 import com.tamnara.backend.poll.util.PollTestBuilder;
+import com.tamnara.backend.user.domain.User;
 import com.tamnara.backend.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -27,15 +29,14 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
-import static com.tamnara.backend.poll.constant.PollResponseMessage.MIN_CHOICES_EXCEED_MAX;
-import static com.tamnara.backend.poll.constant.PollResponseMessage.POLL_NOT_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -60,16 +61,53 @@ class PollServiceImplTest {
 
     @InjectMocks private PollServiceImpl pollServiceImpl;
 
+    private User user;
     private Poll poll;
-    private PollOption option;
+    private PollOption option1;
+    private PollOption option2;
+    private PollOption option3;
     private Poll pollToDelete;
     private Poll pollToPublish;
 
     @BeforeEach
     void setUp() {
-        poll = PollTestBuilder.defaultPoll();
-        ReflectionTestUtils.setField(poll, "id", 1L);
-        option = PollOptionTestBuilder.defaultOption(poll);
+        user = User.builder()
+                .id(1L)
+                .username("testUser")
+                .build();
+
+        poll = Poll.builder()
+                .id(1L)
+                .title("투표 제목")
+                .minChoices(1)
+                .maxChoices(2)
+                .startAt(LocalDateTime.now().minusDays(3))
+                .endAt(LocalDateTime.now().plusDays(3))
+                .state(PollState.PUBLISHED)
+                .options(new ArrayList<>())
+                .build();
+
+        option1 = PollOption.builder()
+                .id(101L)
+                .title("Option 1")
+                .poll(poll)
+                .build();
+
+        option2 = PollOption.builder()
+                .id(102L)
+                .title("Option 2")
+                .poll(poll)
+                .build();
+
+        option3 = PollOption.builder()
+                .id(103L)
+                .title("Option 3")
+                .poll(poll)
+                .build();
+
+        poll.getOptions().add(option1);
+        poll.getOptions().add(option2);
+        poll.getOptions().add(option3);
 
         pollToDelete = Poll.builder()
                 .id(1L)
@@ -102,7 +140,7 @@ class PollServiceImplTest {
                 ));
 
         when(pollRepository.save(any(Poll.class))).thenReturn(poll);
-        when(pollOptionRepository.saveAll(Mockito.anyList())).thenReturn(Arrays.asList(option));
+        when(pollOptionRepository.saveAll(Mockito.anyList())).thenReturn(List.of(option1));
 
         // when
         Long pollId = pollServiceImpl.createPoll(request);
@@ -127,7 +165,7 @@ class PollServiceImplTest {
         // when & then
         assertThatThrownBy(() -> pollServiceImpl.createPoll(request))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(MIN_CHOICES_EXCEED_MAX);
+                .hasMessage(PollResponseMessage.MIN_CHOICES_EXCEED_MAX);
     }
 
     @Test
@@ -137,7 +175,7 @@ class PollServiceImplTest {
         when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.ofNullable(poll));
 
         // when
-        PollInfoResponse response = pollServiceImpl.getLatestPublishedPoll(1L);
+        PollInfoResponse response = pollServiceImpl.getLatestPublishedPoll(user.getId());
 
         // then
         assertEquals(response.getPoll().getId(), poll.getId());
@@ -156,7 +194,7 @@ class PollServiceImplTest {
 
         // then
         assertEquals(HttpStatus.NOT_FOUND, exception.getStatusCode());
-        assertEquals(POLL_NOT_FOUND, exception.getReason());
+        assertEquals(PollResponseMessage.POLL_NOT_FOUND, exception.getReason());
     }
 
     @Test
@@ -236,5 +274,105 @@ class PollServiceImplTest {
         assertEquals(String.format(AlarmMessage.POLL_START_CONTENT, pollToPublish.getTitle()), event.getContent());
         assertEquals(AlarmType.POLLS, event.getTargetType());
         assertNull(event.getTargetId());
+    }
+
+    @Test
+    @DisplayName("투표 기록 저장에 성공")
+    void vote_success() {
+        // given
+        VoteRequest voteRequest = new VoteRequest(List.of(option1.getId(), option2.getId()));
+
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.of(poll));
+        when(pollOptionRepository.findAllById(any())).thenReturn(List.of(option1, option2));
+        when(voteRepository.hasVotedLatestPublishedPoll(user.getId())).thenReturn(false);
+
+        // when
+        pollServiceImpl.vote(user, voteRequest);
+
+        // then
+        verify(voteRepository, times(1)).saveAll(any());
+    }
+
+    @Test
+    @DisplayName("투표 대상 투표가 존재하지 않는 경우 404 예외를 반환한다")
+    void vote_pollNotFound() {
+        // given
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.empty());
+
+        // when
+        ResponseStatusException exception = assertThrows(ResponseStatusException.class, () -> {
+            pollServiceImpl.vote(user, new VoteRequest());
+        });
+
+        // then
+        assertEquals(HttpStatus.NOT_FOUND, exception.getStatusCode());
+        assertEquals(PollResponseMessage.POLL_NOT_FOUND, exception.getReason());
+    }
+
+    @Test
+    @DisplayName("투표가 PUBLISHED 상태가 아닌 경우 403 예외를 반환한다")
+    void vote_pollNotPublished() {
+        // given
+        poll.changeState(PollState.SCHEDULED);
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.of(poll));
+
+        // when & then
+        assertThrows(ResponseStatusException.class, () -> {
+            pollServiceImpl.vote(user, new VoteRequest());
+        });
+    }
+
+    @Test
+    @DisplayName("투표 기간이 아닌 경우 403 예외를 반환한다")
+    void vote_outOfPeriod() {
+        // given
+        LocalDateTime now = LocalDateTime.now();
+        poll = PollTestBuilder.build("test", 1, 2, now.plusDays(1), now.plusDays(3), PollState.PUBLISHED);
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.of(poll));
+
+        // when
+        ResponseStatusException exception = assertThrows(ResponseStatusException.class, () -> {
+            pollServiceImpl.vote(user, new VoteRequest());
+        });
+
+        // then
+        assertEquals(HttpStatus.FORBIDDEN, exception.getStatusCode());
+        assertEquals(PollResponseMessage.POLL_NOT_IN_VOTING_PERIOD, exception.getReason());
+    }
+
+    @Test
+    @DisplayName("이미 투표한 경우 409 예외를 반환한다")
+    void vote_alreadyVoted() {
+        // given
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.of(poll));
+        when(pollOptionRepository.findAllById(List.of(101L))).thenReturn(List.of(option1));
+        when(voteRepository.hasVotedLatestPublishedPoll(user.getId())).thenReturn(true);
+
+        // when
+        ResponseStatusException exception = assertThrows(ResponseStatusException.class, () -> {
+            pollServiceImpl.vote(user, new VoteRequest(List.of(101L)));
+        });
+
+        // then
+        assertEquals(HttpStatus.CONFLICT, exception.getStatusCode());
+        assertEquals(PollResponseMessage.POLL_ALREADY_VOTED, exception.getReason());
+    }
+
+    @Test
+    @DisplayName("선택된 옵션 수가 범위를 벗어나는 경우 400 예외를 반환한다")
+    void vote_optionCountInvalid() {
+        // given
+        when(pollRepository.findLatestPollByPublishedPoll()).thenReturn(Optional.of(poll));
+        when(pollOptionRepository.findAllById(List.of(101L, 102L, 103L)))
+                .thenReturn(List.of(option1, option2, option3));
+
+        // when
+        ResponseStatusException exception = assertThrows(ResponseStatusException.class, () -> {
+            pollServiceImpl.vote(user, new VoteRequest(List.of(101L, 102L, 103L)));
+        });
+
+        // then
+        assertEquals(HttpStatus.BAD_REQUEST, exception.getStatusCode());
+        assertEquals(PollResponseMessage.POLL_INVALID_SELECTION_COUNT, exception.getReason());
     }
 }
